+++ packages/yourloops/lib/data/data-api.ts
@@ -96,9 +96,9 @@ export default class DataApi {
     })
   }
 
-  static async exportData(user: User, patientId: string, startDate: string, endDate: string): Promise<Blob> {
+  static async exportData(user: User, patientId: string, startDate: string, endDate: string): Promise<string> {
     const bgUnits = user.settings?.units ?? Units.gram
-    const { data } = await HttpService.get<Blob>({
+    const { data } = await HttpService.get<string>({
       url: `/export/${patientId}`,
       config: {
         headers: { [HttpHeaderKeys.contentType]: HttpHeaderValues.csv },