+++ CHANGELOG.md
@@ -9,6 +9,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1624 Fix error messages when inviting a patient already in the team
 - YLP-1873 Fix patient profile update not visible in dashboard
 - YLP-1874 Fix PDF report generation
+- YLP-1904 CSV generation outputs a pdf
 
 ## 3.0.1 - 2022-10-11
 ### Added